+++ CHANGELOG.md
@@ -1,9 +1,14 @@
 # YTech Changelog
 
+## []
+
+- Added Potter's Wheel
+- Changed Unfired Clay Bucket recipe
+- Added unfired versions for Flower Pot and Decorated Pot
+
 ## [0.3.3]
 
 - Fixed broken recipes for Thatch and Stone Hammer and changed recipe for grass bed
-
 
 ## [0.3.2]